third_party/blink/web_tests/platform/generic/printing/numberOfPages-expected.txt

Test for internals.numberOfPages()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: number of pages is 2
PASS: number of pages is 1
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
